fix: consolidate mockworkers package to satisfy file-count guard
Merge schema contract constants into project_documentation.go, fold input index case tables into input_index.go, and combine schema topology/metadata/parity tests into schema_contract_test.go so pkg/config/mockworkers stays within the 15-file limit.
